requirements.txt
Flask
boto3
requests
python-dotenv
